Move KeyStoreFilesLocator to be with the only class that uses it in validator api.
