+++ validator/api/src/main/java/tech/pegasys/teku/validator/api/ValidatorConfig.java
@@ -25,7 +25,6 @@ import org.apache.commons.lang3.tuple.Pair;
 import org.apache.logging.log4j.status.StatusLogger;
 import tech.pegasys.teku.bls.BLSPublicKey;
 import tech.pegasys.teku.util.config.InvalidConfigurationException;
-import tech.pegasys.teku.util.config.KeyStoreFilesLocator;
 
 public class ValidatorConfig {
